+++ app/build.gradle
@@ -30,6 +30,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -39,6 +40,7 @@ android {
     }
     buildFeatures {
         compose true
+        buildConfig true
     }
     composeOptions {
         kotlinCompilerExtensionVersion '1.3.2'
@@ -77,6 +79,8 @@ dependencies {
 
     // Compose Navigation
     implementation "androidx.navigation:navigation-compose:2.5.3"
+    // BottomNavigation 및 BottomNavigationItem 구성요소 사용
+    implementation "androidx.compose.material:material:1.4.3"
 
     // Retrofit
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
