--- app/src/main/java/online/partyrun/partyrunapplication/PartyRunApp.kt
@@ -1,13 +1,0 @@
-package online.partyrun.partyrunapplication
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class PartyRunApp: Application() {
-    override fun onCreate() {
-        super.onCreate()
-        Timber.plant(Timber.DebugTree())
-    }
-}
