+++ app/src/main/java/online/partyrun/partyrunapplication/common/Constants.kt
@@ -1,0 +1,8 @@
+package online.partyrun.partyrunapplication.common
+
+import online.partyrun.partyrunapplication.BuildConfig
+
+object Constants {
+    const val BASE_URL = BuildConfig.TEST_BASE_URL
+
+}
