+++ app/src/main/java/online/partyrun/partyrunapplication/data/model/TestQuestionItem.kt
@@ -1,0 +1,8 @@
+package online.partyrun.partyrunapplication.data.model
+
+data class TestQuestionItem(
+    val answer: String,
+    val category: String,
+    val choices: List<String>,
+    val question: String
+)
