+++ app/src/main/java/online/partyrun/partyrunapplication/di/repository/RepositoryModule.kt
@@ -1,0 +1,32 @@
+package online.partyrun.partyrunapplication.di.repository
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import online.partyrun.partyrunapplication.data.repository.TestRepositoryImpl
+import online.partyrun.partyrunapplication.domain.repository.TestRepository
+import javax.inject.Singleton
+
+/*
+    interface인 Repository를 주입하기 위한
+    Repository Module 작성
+*/
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Singleton
+    @Binds
+    abstract fun bindTestRepository (
+        testRepositoryImpl: TestRepositoryImpl
+    ): TestRepository
+
+    /*
+    @Singleton
+    @Binds
+    abstract fun bindAuthRepository (
+        AuthRepositoryImpl: AuthRepositoryImpl
+    ): AuthRepository
+     */
+}
