+++ app/src/main/java/online/partyrun/partyrunapplication/domain/repository/TestRepository.kt
@@ -1,0 +1,9 @@
+package online.partyrun.partyrunapplication.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import online.partyrun.partyrunapplication.network.ApiResponse
+import online.partyrun.partyrunapplication.data.model.TestQuestionItem
+
+interface TestRepository {
+    suspend fun getAllQuestions(): Flow<ApiResponse<List<TestQuestionItem>>>
+}
