--- app/src/main/java/online/partyrun/partyrunapplication/network/api/AuthApiService.kt
@@ -1,4 +1,0 @@
-package online.partyrun.partyrunapplication.network.api
-
-interface AuthApiService {
-}
