--- app/src/main/java/online/partyrun/partyrunapplication/network/models/mockTest/QuestionItem.kt
@@ -1,8 +1,0 @@
-package online.partyrun.partyrunapplication.network.models.mockTest
-
-data class QuestionItem(
-    val answer: String,
-    val category: String,
-    val choices: List<String>,
-    val question: String
-)
