+++ app/src/main/java/online/partyrun/partyrunapplication/network/service/AuthApiService.kt
@@ -1,0 +1,4 @@
+package online.partyrun.partyrunapplication.network.service
+
+interface AuthApiService {
+}
