+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/PartyRunApp.kt
@@ -1,0 +1,13 @@
+package online.partyrun.partyrunapplication.presentation
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class PartyRunApp: Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
+}
