+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/screens/signin/SignInState.kt
@@ -1,0 +1,5 @@
+package online.partyrun.partyrunapplication.presentation.screens.signin
+
+data class SignInState(
+    val test: String
+)
