+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/screens/test/TestState.kt
@@ -1,0 +1,7 @@
+package online.partyrun.partyrunapplication.presentation.screens.test
+
+import online.partyrun.partyrunapplication.data.model.TestQuestionItem
+
+data class TestState(
+    val dataSet: List<TestQuestionItem> = emptyList()
+)
