--- app/src/main/java/online/partyrun/partyrunapplication/util/Constants.kt
@@ -1,6 +1,0 @@
-package online.partyrun.partyrunapplication.util
-
-object Constants {
-    const val BASE_URL = "https://raw.githubusercontent.com/itmmckernan/triviaJSON/master/"
-
-}
